NO JIRA. Fixed URL path percent encoding.
The URLs for easy-auth-info and easy-bag-store were not properly percent-encoded. Do not use
URLEncoder, but use Google guava instead. Guava implements the correct algorithm from the URI
specs. URLEncoder does form encoding, if I remember correctly, which is slightly different.

Also, do not encode the slashes in the path but only the path components themselves.

Also removed the message in the Forbidden response, because it contained the internal bag store
URL, and besides the client know what it has requested, so it is not really necessary to echo
this back.